pyproject.toml
# export SYSTEM_VERSION_COMPAT=1

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sae204"
version = "1.0.0"
description = "SAE204"
dependencies = [
        'click',
        'mysql-connector-python',
        'pandas',
        'numpy',
        'python-dotenv',
        'cachetools'
]
requires-python = ">=3.11"

[project.scripts]
sae204 = "main:cli"

[tool.distutils.bdist_wheel]
universal = true
